state problem fixed + added detail page
when there are objects in an array it's like we are just referencing it rather than not coping it. so made a copy for it. Also created a detail page structute

Furtherly have to make change on the add to cart button using styled component. also the function which is beign called has to be developed to make it re direct to the correct product which has to be shown on the detail page